adxl367: Add temp_en, read acceleration+temperature in one go
For the ADXL367 accelerometer, introduce a DTS setting `temp_en`
to enable/disable temperature readout. Read acceleration and
temperature data with one read command instead of two. This
fixes an issue where the sample rate is less than half of what
is configured as temperature / acceleration data is
alternatingly dropped.

Signed-off-by: Niklas Gürtler <niklas.guertler@e-obs.de>